+++ src/player.rs
@@ -11,11 +11,11 @@ pub struct Player {
     try_sprinting: bool,
     is_sprinting: bool,
     moving_towards: Option<Coordinate>,
-    animation_time: f64,
+    animation_time: f32,
     facing: Direction,
     current_leg: Leg,
     try_walking: Option<Direction>,
-    rotation_timer: f64,
+    rotation_timer: f32,
 }
 
 impl Humanoid for Player {
@@ -67,18 +67,18 @@ impl Humanoid for Player {
     fn set_try_walking(&mut self, try_walking: Option<Direction>) {
         self.try_walking = try_walking;
     }
-    fn get_rotation_timer(&self) -> f64 {
+    fn get_rotation_timer(&self) -> f32 {
         self.rotation_timer
     }
-    fn set_rotation_timer(&mut self, time: f64) {
+    fn set_rotation_timer(&mut self, time: f32) {
         self.rotation_timer = time;
     }
 
-    fn get_animation_time(&self) -> f64 {
+    fn get_animation_time(&self) -> f32 {
         self.animation_time
     }
 
-    fn set_animation_time(&mut self, time: f64) {
+    fn set_animation_time(&mut self, time: f32) {
         self.animation_time = time;
     }
 }
@@ -99,7 +99,7 @@ impl Player {
         }
     }
 
-    pub fn update(&mut self, delta_time: &f64, map: &TileMap, collision_manager: &CollisionManager) {
+    pub fn update(&mut self, delta_time: &f32, map: &TileMap, collision_manager: &CollisionManager) {
         if self.rotation_timer > 0.0 {
             self.rotation_timer -= delta_time;
         }
